Jupiter Notebook file for final project
Attempted to work first with a local python file and upload the entire folder to github, but it's a bit more complicated then expected. 
I have decided to go with the upload of the Jupyter notebook instead, as it's far cleaner.
I do not know how to remove the second graph that is stuck underneath the main one, but technically, the outcome is as expected.